core: allow deterministic date functions in generated columns
