+++ core/translate/index.rs
@@ -1,6 +1,6 @@
 use crate::alloc::TursoIteratorExt;
 use crate::error::SQLITE_CONSTRAINT_UNIQUE;
-use crate::function::{Deterministic, Func, ScalarFunc};
+use crate::function::Func;
 use crate::index_method::IndexMethodConfiguration;
 use crate::numeric::Numeric;
 use crate::schema::{Column, GeneratedType, Table, EXPR_INDEX_SENTINEL, RESERVED_TABLE_PREFIXES};
@@ -22,7 +22,10 @@ use crate::vdbe::builder::{CursorKey, ProgramBuilderOpts, SelfTableContext};
 use crate::vdbe::insn::{to_u16, CmpInsFlags, Cookie};
 use crate::{bail_parse_error, CaptureDataChangesExt, LimboError, MAIN_DB_ID, TEMP_DB_ID};
 use crate::{
-    schema::{BTreeTable, Index, IndexColumn, PseudoCursorType, SchemaObjectType},
+    schema::{
+        is_deterministic_schema_function_call, BTreeTable, Index, IndexColumn, PseudoCursorType,
+        SchemaObjectType,
+    },
     storage::pager::CreateBTreeFlags,
     util::{escape_sql_string_literal, normalize_ident, PRIMARY_KEY_AUTOMATIC_INDEX_NAME_PREFIX},
     vdbe::{
@@ -1018,7 +1021,7 @@ fn validate_index_expression(expr: &Expr, table: &BTreeTable) -> bool {
     let is_deterministic_fn = |name: &str, args: &[Box<Expr>]| {
         let n = normalize_ident(name);
         Func::resolve_function(&n, args.len())
-            .is_ok_and(|f| f.is_some_and(|f| is_valid_index_function_call(&f, args)))
+            .is_ok_and(|f| f.is_some_and(|f| is_deterministic_schema_function_call(&f, args)))
     };
 
     let mut ok = true;
@@ -1087,74 +1090,6 @@ fn validate_index_expression(expr: &Expr, table: &BTreeTable) -> bool {
         })
     });
     ok
-}
-
-fn is_valid_index_function_call(func: &Func, args: &[Box<Expr>]) -> bool {
-    match func {
-        Func::Scalar(
-            ScalarFunc::Date
-            | ScalarFunc::Time
-            | ScalarFunc::DateTime
-            | ScalarFunc::UnixEpoch
-            | ScalarFunc::JulianDay
-            | ScalarFunc::StrfTime
-            | ScalarFunc::TimeDiff,
-        ) => is_deterministic_datetime_index_call(func, args),
-        _ => func.is_deterministic(),
-    }
-}
-
-// SQLite allows date/time functions in expression indexes only when they do not
-// depend on the current clock or local timezone.
-fn is_deterministic_datetime_index_call(func: &Func, args: &[Box<Expr>]) -> bool {
-    match func {
-        Func::Scalar(ScalarFunc::Date)
-        | Func::Scalar(ScalarFunc::Time)
-        | Func::Scalar(ScalarFunc::DateTime)
-        | Func::Scalar(ScalarFunc::UnixEpoch)
-        | Func::Scalar(ScalarFunc::JulianDay) => {
-            !args.is_empty()
-                && !is_current_time_expr(args[0].as_ref())
-                && !args[1..]
-                    .iter()
-                    .any(|arg| is_unsafe_datetime_modifier(arg.as_ref()))
-        }
-        Func::Scalar(ScalarFunc::StrfTime) => {
-            args.len() >= 2
-                && !is_current_time_expr(args[1].as_ref())
-                && !args[2..]
-                    .iter()
-                    .any(|arg| is_unsafe_datetime_modifier(arg.as_ref()))
-        }
-        Func::Scalar(ScalarFunc::TimeDiff) => {
-            !args.iter().any(|arg| is_current_time_expr(arg.as_ref()))
-        }
-        _ => unreachable!("non-datetime function passed to datetime index validator"),
-    }
-}
-
-fn is_current_time_expr(expr: &Expr) -> bool {
-    matches!(
-        expr,
-        Expr::Literal(ast::Literal::String(value)) if string_literal_eq(value, "now")
-    ) || matches!(
-        expr,
-        Expr::Literal(
-            ast::Literal::CurrentDate | ast::Literal::CurrentTime | ast::Literal::CurrentTimestamp
-        )
-    )
-}
-
-fn is_unsafe_datetime_modifier(expr: &Expr) -> bool {
-    matches!(
-        expr,
-        Expr::Literal(ast::Literal::String(value))
-            if string_literal_eq(value, "localtime") || string_literal_eq(value, "utc")
-    ) || is_current_time_expr(expr)
-}
-
-fn string_literal_eq(value: &str, expected: &str) -> bool {
-    value.trim_matches('\'').eq_ignore_ascii_case(expected)
 }
 
 fn emit_index_column_value_from_cursor(
